requirements.txt
# Core dependencies
numpy>=1.21.0
opencv-python>=4.5.0
Pillow>=8.0.0

# PyTorch and vision
torch>=1.9.0
torchvision>=0.10.0

# Ultralytics for SAM3
ultralytics>=8.0.0
